assigning values, system redone

+++ Assets/Scripts/CrimeGen/commit.cs
@@ -5,6 +5,6 @@ using System.Collections;
 public class commit
 {
     public string committance;
-    public int spectrumValue;
+    public float spectrumValue;
 
 }
